Tweak tag bar square coloring
Originally in dwm the square is simply a filled rectangle or an outline.
But ever since I decreased lrpan all tags are very close together and it
looks ugly when the outline of a rectangle overlaps with a digit or
icon. So I added another draw call and basically now there's always a
filled rectangle area above the tag characters and below the outline.

Also added some missing stuff in config.def.h