Change float center map from m to n
The main reason is that one of my laptops' keyboards doesn't allow
WIN+ALT+SHIFT+M to be pressed at all due to hardware limitations,
but WIN+ALT+SHIFT+N works. At first I had a separate config for that
computer, but then I noticed that N is actually easier to reach in that
position and feels more comfortable, so I'm making this into an official
change.